requirements.txt
django~=5.2.4
djangorestframework~=3.16.0
django-filter~=25.1
drf_spectacular
django-cors-headers
psycopg
Pillow